requirements.txt
flask>=2.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
openai>=1.0.0
pypdf>=3.0.0
pdfminer.six>=20221105
python-dotenv>=1.0.0
